Moved cast.py under triton dir, made changes everywhere to use quantize_triton or tex.quantize based on env variable
